Upgraded to FontAwesome 5. Had to update every FA icon since they switched terminology. Some icons also changed names or were eliminated in the process; those have been altered to match or replacement icons found.
Updated how Locations are populated on Event searches.
Old implementation, which was to use the relation defined on the Event, was unexpectedly slow to populate.
This was changed to not getting that relation, but rather the location_id_string, then making a separate call to get the Locations by ID.
This solution is notably faster, but at some point it should be investigated further why the old solution was so slow.